Audit round 2: 30 confirmed findings fixed, incl. production drift
Round 2 attacked round 1's own fix commit. It was not dry, and the
sharpest finding was about deployment, not code:

- The live Wheeler node was still executing PRE-FIX consensus code. Round
  1 moved the gates into an isolated tree (build/ci) so they could never
  disturb production, which also meant a green gate proved nothing about
  the running node: the 64-bit KawPow DAG fix was verified in a tree the
  node does not use. Deployment is now an explicit, verified act
  (scripts/deploy_node.sh: prep, build, assert binary matches repo
  sources, restart, prove the chain continues, audit). Production is on
  the fixed binary; chain continued 3965 -> 3966, audit wei-exact.
- verify_devnet.sh's new London gate was a SyntaxError on every Python
  (backslash inside an f-string expression). It survived a full audit
  round because NO gate ran that script. ci_devnet.sh now runs it, so the
  operator flow is covered by the e2e gate.
- ship_stratum_e2e.sh: over-escaped $env:USERPROFILE made every pc3080
  metric read 0/empty, i.e. the report fabricated the ideal result
  (proven against the live 11,880-line miner log: 4,015 accepted shares
  where the old form returned 0). kawpowminer never logs Speed at all, so
  that row is now a computed hashrate. The script also boots its own
  KawPow devnet (the documented precondition was unobtainable), proves
  the algorithm from the node log rather than assuming it, checks the
  scp, and FAILS on zero blocks or zero accepted shares.
- Same uint32 wrap class as round 1, in kawpowGenerateDataset's word
  offset; ci_prepare's self-heal could rm -rf a tree a live node runs
  from; run-node.sh derived the wrong networkid for absolute-path genesis
  and its Art VIII guard was line-based (multi-line JSON evaded it,
  decimal-prefix IDs falsely blocked).
- Article VIII had no guard on the distribution channel: NET=everett
  produced a tarball that started the reserved mainnet unconditionally.
  Guard added to the dist runner and to make_dist.sh, which now also
  refuses to package an unpinned, unpatched, or stale binary. Deleted two
  stale pre-KawPow staged packages (untracked), one of them a runnable
  15537393 node.
- Dashboard: reorg handling drops blocks above a decreased head and
  sweeps the window on divergence; the audit reaper only kills orphans
  (ppid 1) and verifies group leadership; the GPU/CPU origin label no
  longer inverts after 16 miner reconnects.
- burn_audit's PASS banner no longer claims "every account exact" when it
  skipped contract-touched accounts.
- Docs: ethminer-as-default topology corrected repo-wide, EVERETT_KAWPOW
  documented, GPU_MINING no longer teaches --http.vhosts '*', RUNBOOK's
  negative-control claim scoped to the gates actually controlled,
  GENESIS_SPEC section 4 corrected to London-terminal.

Verified: consensus gates green, hermetic e2e PASSED (now including
verify_devnet.sh), consistency gate 13/13, live audit wei-exact @3970,
dashboard healthy, stratum race tests green, zero em dashes.

+++ client/kawpow_engine.go
@@ -149,7 +149,11 @@ func kawpowGenerateDataset(dest []uint32, cache []uint32) {
 			}
 			for index := first; index < limit; index++ {
 				item := kawpowGenerateDatasetItem(cache, index, keccak512)
-				copy(dest[index*hashWords:], asUint32Slice(item))
+				// uint64 for the word offset: index is a 64-byte item
+				// number, so index*hashWords crosses uint32 at a 16 GiB
+				// DAG (epoch 1921), the same wrap the loop's DAG offsets
+				// had. Item numbers themselves stay exact to 256 GiB.
+				copy(dest[uint64(index)*hashWords:], asUint32Slice(item))
 			}
 		}(uint32(i))
 	}
